refactor code and update unit tests
